requirements.txt
google-genai==1.0.0
mysql-connector-python==8.0.33
requests
numpy
soundfile
pydub
python-dotenv
scipy
silero-vad
onnxruntime
